+++ tests/test_interface.py
@@ -1,6 +1,7 @@
 import pytest
 from flaky import flaky
 import os
+from ast import literal_eval
 from base64 import b64encode, b64decode
 from selenium import webdriver
 from selenium.webdriver.support.ui import WebDriverWait
@@ -25,19 +26,42 @@ from mols2grid.select import register
 geckodriver_autoinstaller.install()
 pytestmark = pytest.mark.webdriver
 
+HEADLESS = True
+
+class selection_available:
+    def __init__(self, is_empty=False):
+        self.empty = is_empty
+
+    def __call__(self, driver):
+        sel = driver.execute_script("return SELECTION.to_dict();")
+        sel = literal_eval(sel)
+        if sel == {} and self.empty:
+            return True
+        elif sel != {} and not self.empty:
+            return sel
+        return False
+
 class FirefoxDriver(webdriver.Firefox):
-    def wait_for_img_load(self, max_delay=15, selector="#mols2grid .cell .data-img svg"):
+    def wait_for_img_load(self, max_delay=5, selector="#mols2grid .cell .data-img svg"):
         return (WebDriverWait(self, max_delay)
-                .until(EC.presence_of_element_located(
-                        (By.CSS_SELECTOR, selector))))
+                .until(EC.presence_of_element_located((By.CSS_SELECTOR,
+                                                       selector))))
+
+    def wait_for_selection(self, is_empty=False, max_delay=3):
+        return (WebDriverWait(self, max_delay)
+                .until(selection_available(is_empty)))
 
     def wait(self, condition, max_delay=5):
         return (WebDriverWait(self, max_delay,
-                      ignored_exceptions=[StaleElementReferenceException])
+                ignored_exceptions=[StaleElementReferenceException])
                 .until(condition))
 
     def find_by_id(self, element_id, **kwargs):
         condition = EC.presence_of_element_located((By.ID, element_id))
+        return self.wait(condition, **kwargs)
+
+    def find_clickable(self, by, selector, **kwargs):
+        condition = EC.element_to_be_clickable((by, selector))
         return self.wait(condition, **kwargs)
 
     def find_by_css_selector(self, css_selector, **kwargs):
@@ -63,16 +87,27 @@ class FirefoxDriver(webdriver.Firefox):
         im = Image.open(BytesIO(b64decode(img.get_attribute("src")[22:])))
         return md5(im.tobytes()).hexdigest()
 
+    def substructure_query(self, smarts):
+        self.find_clickable(By.ID, "searchBtn").click()
+        self.find_clickable(By.ID, "smartsSearch").click()
+        self.find_by_id("searchbar").send_keys(smarts)
+        self.wait_for_img_load()
+
+    def text_search(self, txt):
+        self.find_clickable(By.ID, "searchBtn").click()
+        self.find_clickable(By.ID, "txtSearch").click()
+        self.find_by_id("searchbar").send_keys(txt)
+        self.wait_for_img_load()
 
 def determine_scope(fixture_name, config):
-    if os.environ.get("GITHUB_ACTIONS", False):
+    if os.environ.get("GITHUB_ACTIONS"):
         return "function"
     return "module"
 
 @pytest.fixture(scope=determine_scope)
 def driver():
     options = webdriver.FirefoxOptions()
-    options.headless = True
+    options.headless = HEADLESS
     driver = FirefoxDriver(options=options)
     driver.set_page_load_timeout(10)
     yield driver
@@ -108,7 +143,7 @@ def html_doc(grid):
 # make sure non-parametrized test is ran first
 @pytest.mark.order(1)
 def test_no_subset_all_visible(driver, grid):
-    doc = get_doc(grid, {"selection": False})
+    doc = get_doc(grid, {"tooltip": [], "selection": False})
     driver.get(doc)
     columns = set(grid.dataframe.columns.drop("mol").to_list())
     cell = driver.find_by_css_selector("#mols2grid .cell")
@@ -129,7 +164,8 @@ def test_smiles_hidden(driver, html_doc):
 @pytest.mark.parametrize("n_cols", [1, 3])
 @pytest.mark.parametrize("n_rows", [1, 3])
 def test_page_click(driver, grid, page, n_cols, n_rows):
-    doc = get_doc(grid, dict(n_cols=n_cols, n_rows=n_rows))
+    doc = get_doc(grid, dict(subset=["img", "_Name"],
+                             n_cols=n_cols, n_rows=n_rows))
     driver.get(doc)
     for i in range(2, page+1):
         driver.wait_for_img_load()
@@ -160,11 +196,7 @@ def test_css_properties(driver, grid, name, css_prop, value, expected):
 def test_text_search(driver, html_doc):
     driver.get(html_doc)
     driver.wait_for_img_load()
-    (ActionChains(driver)
-        .send_keys_to_element(driver.find_by_id("searchbar"), "iodopropane")
-        .key_up(Keys.TAB)
-        .pause(.5)
-        .perform())
+    driver.text_search("iodopropane")
     el = driver.find_by_css_selector("#mols2grid .cell .data-SMILES")
     assert el.get_attribute("innerHTML") == "CC(I)C"
 
@@ -172,26 +204,15 @@ def test_text_search(driver, html_doc):
 def test_smarts_search(driver, html_doc):
     driver.get(html_doc)
     driver.wait_for_img_load()
-    (ActionChains(driver)
-        .click(driver.find_by_id("searchBtn"))
-        .pause(.5)
-        .click(driver.find_by_id("smartsSearch"))
-        .send_keys_to_element(driver.find_by_id("searchbar"), "CC(I)C")
-        .key_up(Keys.TAB)
-        .pause(.5)
-        .perform())
+    driver.substructure_query("CC(I)C")
     el = driver.find_by_css_selector("#mols2grid .cell .data-_Name")
     assert el.text == "2-iodopropane"
 
 def test_selection_click(driver, html_doc):
     driver.get(html_doc)
     driver.wait_for_img_load()
-    (ActionChains(driver)
-        .click(driver.find_by_css_selector("input[type='checkbox']"))
-        .pause(.5)
-        .perform())
-    sel = driver.execute_script("return SELECTION.to_dict();")
-    sel = eval(sel)
+    driver.find_clickable(By.CSS_SELECTOR, "input[type='checkbox']").click()
+    sel = driver.wait_for_selection(is_empty=False)
     assert sel == {0: "CCC(C)CC"}
     register._clear()
 
@@ -202,66 +223,38 @@ def test_selection_with_cache_check_and_uncheck(driver, df):
     doc = get_doc(grid, {})
     driver.get(doc)
     driver.wait_for_img_load()
-    sel = driver.execute_script("return SELECTION.to_dict();")
-    sel = eval(sel)
+    sel = driver.wait_for_selection(is_empty=False)
     assert sel == {0: "CCC(C)CC"}
-    (ActionChains(driver)
-        .click(driver.find_by_css_selector("input[type='checkbox']"))
-        .pause(.5)
-        .perform())
-    sel = driver.execute_script("return SELECTION.to_dict();")
-    sel = eval(sel)
-    assert sel == {}
+    driver.find_clickable(By.CSS_SELECTOR, "input[type='checkbox']").click()
+    empty_sel = driver.wait_for_selection(is_empty=True)
+    assert empty_sel is True
     register._clear()
 
 def test_selection_check_uncheck_invert(driver, html_doc):
     driver.get(html_doc)
     driver.wait_for_img_load()
-    (ActionChains(driver)
-        .click(driver.find_by_id("chkboxDropdown"))
-        .pause(.5)
-        .click(driver.find_by_id("btn-chkbox-all"))
-        .perform())
-    sel = driver.execute_script("return SELECTION.to_dict();")
-    sel = eval(sel)
+    # search
+    driver.text_search("iodopropane")
+    # check all
+    driver.find_clickable(By.ID, "chkboxDropdown").click()
+    driver.find_clickable(By.ID, "btn-chkbox-all").click()
+    sel = driver.wait_for_selection(is_empty=False)
     assert len(sel) == 30
-    (ActionChains(driver)
-        .click(driver.find_by_id("chkboxDropdown"))
-        .pause(.5)
-        .click(driver.find_by_id("btn-chkbox-none"))
-        .perform())
-    sel = driver.execute_script("return SELECTION.to_dict();")
-    sel = eval(sel)
-    assert sel == {}
-    (ActionChains(driver)
-        .click(driver.find_by_css_selector("input[type='checkbox']"))
-        .click(driver.find_by_id("chkboxDropdown"))
-        .pause(.5)
-        .click(driver.find_by_id("btn-chkbox-invert"))
-        .perform())
-    sel = driver.execute_script("return SELECTION.to_dict();")
-    sel = eval(sel)
-    assert len(sel) == 29
-    register._clear()
-
-def test_check_all_selects_only_not_hidden(driver, html_doc):
-    driver.get(html_doc)
-    driver.wait_for_img_load()
-    (ActionChains(driver)
-        .send_keys_to_element(driver.find_by_id("searchbar"), "iodopropane")
-        .key_up(Keys.TAB)
-        .click(driver.find_by_id("chkboxDropdown"))
-        .pause(.3)
-        .click(driver.find_by_id("btn-chkbox-all"))
-        .perform())
-    driver.find_by_id("searchbar").clear()
-    (ActionChains(driver)
-        .send_keys_to_element(driver.find_by_id("searchbar"), Keys.BACKSPACE)
-        .key_up(Keys.TAB)
-        .perform())
-    sel = driver.execute_script("return SELECTION.to_dict();")
-    sel = eval(sel)
+    # uncheck all
+    driver.find_clickable(By.ID, "chkboxDropdown").click()
+    driver.find_clickable(By.ID, "btn-chkbox-none").click()
+    empty_sel = driver.wait_for_selection(is_empty=True)
+    assert empty_sel is True
+    # check matching
+    driver.find_clickable(By.ID, "chkboxDropdown").click()
+    driver.find_clickable(By.ID, "btn-chkbox-match").click()
+    sel = driver.wait_for_selection(is_empty=False)
     assert sel == {27: 'CC(I)C'}
+    # invert
+    driver.find_clickable(By.ID, "chkboxDropdown").click()
+    driver.find_clickable(By.ID, "btn-chkbox-invert").click()
+    sel = driver.wait_for_selection(is_empty=False)
+    assert len(sel) == 29
     register._clear()
 
 @pytest.mark.parametrize("prerender", [True, False])
@@ -285,7 +278,7 @@ def test_image_use_coords(driver, df):
     if rdkit_version == "2020.03.1":
         assert md5_hash == "aed60ed28347831d24f02dbb5be19007"
     else:
-        assert md5_hash == "6909ba43f86003cea9b0fd8d723cddfe"
+        assert md5_hash == "80ac785821c6aee7923c793930c4723e"
 
 @pytest.mark.parametrize(["coordGen", "prerender", "expected"], [
     (True,  True, "acf7cf7cd5cfaa5eaf4a4f257e290e49"),
@@ -311,9 +304,9 @@ def test_coordgen(driver, mols, coordGen, prerender, expected):
     (True, True, "acf7cf7cd5cfaa5eaf4a4f257e290e49"),
     (False, True, ("2d30379732d312409d1d93ad5b3c3e60"
                    if rdkit_version == "2020.03.1"
-                   else "b01c2146523b8a006bb34621839668eb")),
+                   else "b163a9cb19caee002ff9d739ed513672")),
     (True, False, "d8227a9948b6f801193085e5a8b257a2"),
-    (False, False, "e7b837b4d0c5f1ed8e1cd79734b6845e"),
+    (False, False, "a8d830b6127c390dd5759a6ab1af8b3a"),
 ])
 def test_removeHs(driver, df, removeHs, prerender, expected):
     useSVG = not prerender
@@ -333,10 +326,10 @@ def test_removeHs(driver, df, removeHs, prerender, expected):
     assert md5_hash == expected
 
 @pytest.mark.parametrize(["kwargs", "expected"], [
-    (dict(addAtomIndices=True), "d9611dd45ff82ab73cf31dfc3a33706d"),
+    (dict(addAtomIndices=True), "e01cd9295b091e9e2167568bd3cf139d"),
     (dict(fixedBondLength=10), "1dfa2a5c43022ce66a2ff26a517698cf"),
     (dict(atomColourPalette={6: (0, .8, .8)}), "91d114d421fc6a05e4537d19ad4b8324"),
-    (dict(legend="foo"), "bb5bc6ae713cfaea4e6374fc35e4be4e"),
+    (dict(legend="foo"), "49cfe8038b8fc0b0401fd0056426d6b7"),
 ])
 def test_moldrawoptions(driver, df, kwargs, expected):
     grid = get_grid(df, **kwargs)
@@ -377,9 +370,7 @@ def test_tooltip_trigger(driver, grid):
         .perform())
     with pytest.raises(NoSuchElementException):
         driver.find_element_by_css_selector('div.popover[role="tooltip"]')
-    (ActionChains(driver)
-        .click(driver.find_by_css_selector("#mols2grid .cell .data-img *"))
-        .perform())
+    driver.find_clickable(By.CSS_SELECTOR, "#mols2grid .cell .data-img *").click()
     tooltip = driver.find_by_css_selector('div.popover[role="tooltip"]')
     el = tooltip.find_element_by_class_name("popover-body")
     assert el.get_attribute("innerHTML") == "<strong>_Name</strong>: 3-methylpentane"
@@ -472,20 +463,18 @@ def test_transform_style_tooltip(driver, grid):
 @pytest.mark.parametrize("selection", [True, False])
 def test_callback_js(driver, grid, selection):
     doc = get_doc(grid, {
+        "subset": ["img", "_Name"],
         "callback": "$('#mols2grid .cell .data-_Name').html('foo')",
         "selection": selection
     })
     driver.get(doc)
     driver.wait_for_img_load()
-    (ActionChains(driver)
-        .click(driver.find_by_css_selector("#mols2grid .cell .data-img"))
-        .pause(.5)
-        .perform())
+    driver.find_clickable(By.CSS_SELECTOR, "#mols2grid .cell .data-img").click()
     el = driver.find_by_css_selector("#mols2grid .cell .data-_Name")
     assert el.text == "foo"
 
 def test_sort_by(driver, grid):
-    doc = get_doc(grid, {"sort_by": "_Name"})
+    doc = get_doc(grid, {"subset": ["img", "_Name"], "sort_by": "_Name"})
     driver.get(doc)
     el = driver.find_by_css_selector("#mols2grid .cell .data-_Name")
     assert el.text == "1,1,2,2-tetrachloroethane"
@@ -493,37 +482,26 @@ def test_sort_by(driver, grid):
 def test_sort_button(driver, html_doc):
     driver.get(html_doc)
     driver.wait_for_img_load()
-    (ActionChains(driver)
-        .click(driver.find_by_id("sortDropdown"))
-        .pause(.3)
-        .click(driver.find_by_css_selector('button.sort-btn[data-name="data-_Name"]'))
-        .perform())
+    driver.find_clickable(By.ID, "sortDropdown").click()
+    driver.find_clickable(By.CSS_SELECTOR,
+                          'button.sort-btn[data-name="data-_Name"]').click()
     el = driver.find_by_css_selector("#mols2grid .cell .data-_Name")
     assert el.text == "1,1,2,2-tetrachloroethane"
-    (ActionChains(driver)
-        .click(driver.find_by_id("sortDropdown"))
-        .pause(.3)
-        .click(driver.find_by_css_selector('button.sort-btn[data-name="data-_Name"]'))
-        .perform())
+    driver.find_clickable(By.ID, "sortDropdown").click()
+    driver.find_clickable(By.CSS_SELECTOR,
+                          'button.sort-btn[data-name="data-_Name"]').click()
     el = driver.find_by_css_selector("#mols2grid .cell .data-_Name")
     assert el.text == "tetrachloromethane"
 
 @pytest.mark.parametrize(["substruct_highlight", "expected"], [
-    (True,  "6a1ae3fc97cbe003c86db3c624e5ff09"),
-    (False, "4884c75977e8c3e0fa2e6a1607275406")
+    (True,  "2d943c1c2ba5e4bc75ed307d9e1dc456"),
+    (False, "343b129e85fd56945a7bcfa7c12d63f7")
 ])
 def test_substruct_highlight(driver, grid, substruct_highlight, expected):
     doc = get_doc(grid, {"n_rows": 1, "substruct_highlight": substruct_highlight})
     driver.get(doc)
-    text_box = driver.find_by_id("searchbar")
     driver.wait_for_img_load()
-    (ActionChains(driver)
-        .click(driver.find_by_id("searchBtn"))
-        .pause(.5)
-        .click(driver.find_by_id("smartsSearch"))
-        .send_keys_to_element(text_box, "CC(I)C")
-        .key_up(Keys.TAB)
-        .perform())
+    driver.substructure_query("CC(I)C")
     md5_hash = driver.get_svg_md5_hash()
     assert md5_hash == expected
 
@@ -531,19 +509,11 @@ def test_substruct_clear_removes_highlight(driver, grid):
     doc = get_doc(grid, {"n_rows": 1, "substruct_highlight": True})
     driver.get(doc)
     driver.wait_for_img_load()
-    (ActionChains(driver)
-        .click(driver.find_by_id("searchBtn"))
-        .pause(.5)
-        .click(driver.find_by_id("smartsSearch"))
-        .send_keys_to_element(driver.find_by_id("searchbar"), "C")
-        .key_up(Keys.TAB)
-        .perform())
+    driver.substructure_query("C")
     md5_hash_hl = driver.get_svg_md5_hash()
     driver.find_by_id("searchbar").clear()
-    (ActionChains(driver)
-        .send_keys_to_element(driver.find_by_id("searchbar"), Keys.BACKSPACE)
-        .key_up(Keys.TAB)
-        .perform())
+    driver.find_by_id("searchbar").send_keys(Keys.BACKSPACE)
+    driver.wait_for_img_load()
     md5_hash = driver.get_svg_md5_hash()
     assert md5_hash != md5_hash_hl
     assert md5_hash == "d8227a9948b6f801193085e5a8b257a2"
@@ -552,27 +522,15 @@ def test_smarts_to_text_search_removes_highlight(driver, grid):
     doc = get_doc(grid, {"n_rows": 1, "substruct_highlight": True})
     driver.get(doc)
     driver.wait_for_img_load()
-    (ActionChains(driver)
-        .click(driver.find_by_id("searchBtn"))
-        .pause(.5)
-        .click(driver.find_by_id("smartsSearch"))
-        .send_keys_to_element(driver.find_by_id("searchbar"), "I")
-        .key_up(Keys.TAB)
-        .perform())
+    driver.substructure_query("I")
     md5_hash_hl = driver.get_svg_md5_hash()
-    (ActionChains(driver)
-        .click(driver.find_by_id("searchBtn"))
-        .pause(.5)
-        .click(driver.find_by_id("txtSearch"))
-        .send_keys_to_element(driver.find_by_id("searchbar"), "odopropane")
-        .key_up(Keys.TAB)
-        .perform())
+    driver.text_search("odopropane")
     md5_hash = driver.get_svg_md5_hash()
     assert md5_hash != md5_hash_hl
-    assert md5_hash == "4884c75977e8c3e0fa2e6a1607275406"
+    assert md5_hash == "343b129e85fd56945a7bcfa7c12d63f7"
 
 def test_filter(driver, grid):
-    doc = get_doc(grid, {})
+    doc = get_doc(grid, {"subset": ["img", "_Name"],})
     driver.get(doc)
     mask = grid.dataframe["_Name"].str.contains("iodopropane")
     ids = grid.dataframe.loc[mask]["mols2grid-id"].to_list()
@@ -615,6 +573,7 @@ def test_colname_with_spaces(driver, df):
     el = driver.find_by_css_selector("#mols2grid .cell .data")
     assert el.text == "3-methylpentane"
 
+@flaky(max_runs=3, min_passes=1)
 def test_custom_header(driver, grid):
     doc = get_doc(grid, {
         "subset": ["img"],
@@ -624,3 +583,78 @@ def test_custom_header(driver, grid):
     driver.wait_for_img_load()
     val = driver.execute_script("return RDKit.version();")
     assert val == "2021.09.2"
+
+def test_table_template(driver):
+    df = mols2grid.sdf_to_dataframe(sdf_path)[:15]
+    grid = mols2grid.MolGrid(df, mol_col="mol", prerender=True)
+    doc = get_doc(grid, dict(
+        template="table", 
+        subset=["mols2grid-id", "img"],
+        tooltip=["_Name"],
+        sort_by="_Name"
+    ))
+    driver.get(doc)
+    el = driver.find_by_css_selector("#mols2grid td.col-0")
+    assert el.find_element_by_class_name("data-mols2grid-id").text == "8"
+    (ActionChains(driver)
+        .move_to_element(el.find_element_by_css_selector(".data-img *"))
+        .perform())
+    tooltip = driver.find_by_css_selector('div.popover[role="tooltip"]')
+    el = tooltip.find_element_by_class_name("popover-body")
+    assert el.get_attribute("innerHTML") == "<strong>_Name</strong>: 1,3,5-trimethylbenzene"
+    md5_hash = driver.get_svg_md5_hash("#mols2grid td .data-img")
+    if rdkit_version == "2020.03.1":
+        assert md5_hash == "7ca709f65c41fcfe090e98525920fb40"
+    else:
+        assert md5_hash == "9a681d0d978a6cb7ed28a95850d775b8"
+
+def test_default_subset_tooltip(driver, grid):
+    doc = get_doc(grid, {"n_rows": 1})
+    driver.get(doc)
+    driver.wait_for_img_load()
+    expected_subset = ["mols2grid-id", "img"]
+    expected_tooltip = [x for x in grid.dataframe.columns.drop("mol").to_list()
+                        if x not in expected_subset]
+    cell = driver.find_by_css_selector("#mols2grid .cell")
+    data_elements = cell.find_elements_by_class_name("data")
+    subset = [c.replace("data-", "").replace("-copy", "")
+               for x in data_elements
+               for c in x.get_attribute("class").split(" ")
+               if c.startswith("data-") and not x.get_attribute("style")]
+    assert subset == expected_subset
+    tooltip = [c.replace("data-", "").replace("-copy", "")
+               for x in data_elements
+               for c in x.get_attribute("class").split(" ")
+               if c.startswith("data-")
+               and x.get_attribute("style") == "display: none;"]
+    assert tooltip == expected_tooltip
+
+def test_multi_highlight(driver, html_doc):
+    driver.get(html_doc)
+    driver.wait_for_img_load()
+    driver.substructure_query("Br")
+    md5_hash = driver.get_svg_md5_hash()
+    assert md5_hash == "814a4c5dfa5226b629031861508505c5"
+
+def test_single_highlight(driver, grid):
+    doc = get_doc(grid, {"single_highlight": True})
+    driver.get(doc)
+    driver.wait_for_img_load()
+    driver.substructure_query("Br")
+    md5_hash = driver.get_svg_md5_hash()
+    assert md5_hash == "ca043475667e4f79a3b87a96d7dd731a"
+
+def test_mol_depiction_aligned_to_query(driver, html_doc):
+    driver.get(html_doc)
+    driver.wait_for_img_load()
+    driver.substructure_query("CCCBr")
+    images = driver.wait(EC.presence_of_all_elements_located(
+                         (By.CSS_SELECTOR, "#mols2grid .cell .data-img")))
+    for img, expected in zip(images, [
+        "4b19aa7cf16fe691ea7840a3ce9601b5",
+        "4d7a5d8cd6652c1fd8d04a6fffdd2e3a",
+    ]):
+        im = svg2png(bytestring=(img.get_attribute("innerHTML")))
+        im = Image.open(BytesIO(im))
+        md5_hash = md5(im.tobytes()).hexdigest()
+        assert md5_hash == expected